Fix a bug where prevSpecial(x, false) would fail
Copy+paste error :)
nextSpecial calls itself when it found a special frame but is supposed to skip them.
prevSpecial also called nextSpecial instead of prevSpecial in that case,
leading to getting stuck on one special frame if the previous one was a debated one.
(This was not visible on the website since the boolean parameter was never used.)