Refactor ayop-logo.svg
The only really visible change is that the pyramid (and everything inside)
moves to the top of the image, but behind the scenes, the pyramid was
completely rewritten to remove the huge Inkscape bloat that it used to be.
Now, the internal structure is easily visible, allowing simply changes to every
part of the image. Also, the lines of the bricks got a class (.brick_line), so
that they can easily be styled.