Minor stylesheet changes to padding/margin
Previously, if you collapsed the “Frame Data” panel, a three pixels high
gap would show. This commit fixes that by removing the 3px bottom
padding from ‘.openingpanel’s. This lead to the bottom of the content
inside the panels being to close to the bottom of the panels, so the
-3px margin was removed as well.

Also, the spacing of the ‘moreinfo’ list entries (the links to the
comic, explainxkcd, et.c) has been changed from “10px above each entry”
to “5px vertically around each entry” to fix the same problem. The
spacing between the entries remains the same, but now there is a larger
gap between the last entry and the end of the “More Info” panel.