Fix bug where panels couldn’t be closed
Sometimes, the init function for a panel throws an exception. (For me,
this happens during local testing because my setup isn’t complete.)
Since that function is called from the slideToggle callback, the
exception would be thrown to jQuery, which apparently didn’t like that,
with the effect of subsequent slideToggles no longer working. Fix by
catching (and ignoring) any exceptions in the callback.