Center logo below “at your own pace” h2
This centers the AYOP logo horizontally in regard to the width of the
“at your own pace” h2 element.

The IE≤9 warning was moved because between the h2 and the logo, it would
have caused _these_ to be centered in regards to the warning’s width,
which completely breaks all kinds of design on the page. The element
will still mess up the page, but not quite as bad.